Fix event layout for decoding in perfview

+++ src/Ultra.Sampler/MacOS/NativeCallstackDelegate.cs
@@ -4,4 +4,4 @@
 
 namespace Ultra.Sampler.MacOS;
 
-internal unsafe delegate void NativeCallstackDelegate(ulong threadId, nint pFrames, int frameCount);
+internal unsafe delegate void NativeCallstackDelegate(ulong threadId, ulong pFrames, int frameCount);
